python/s0102/requirements.txt
openai==1.59.3
python-dotenv==1.0.1